Allow configuration to be passed in
Previously it was necessary to edit source + rebuild metrics-play to change the
prefix label that metrics had, or to change which HTTP response codes had
individual metrics.

This change enables the passing in of configuration to override these, whilst
defaulting to the previous behaviour.

It also optionally allows the specifying of healthCheck URIs - enabling health
checks to be classified seperately from regular traffic, to prevent skewing of
metrics from monitoring.