Handle uncaught exceptions in MetricsFilter
In the case of an uncaught exception during a request, logCompleted is
never called. Common practice in Play apps is to have a global onError handler
that returns an HTTP 500 result. Due to logCompleted never being called,
these 500 results do not get caught by the MetricsFilter. Additionally,
their timing is ignored, and activeRequests is never decremented.

This change addresses these issues by calling logCompleted with an
InternalServerError result before propagating the exception. This also
exchanges EssentialFilter for Filter, as EssentialFilter is lower-level
and is not necessary for this class.